Refactor application by restructuring modules and components
Detailed explanation:
The commit includes restructuring of the application where unnecessary individual modules (TodoEditorModule, TodoFormModule, TodoItemModule) were eliminated and replaced by CoreModule and FeaturesModule. It also includes newly added components (HeaderComponent, TodoComponent) and updates for the AppComponent. This restructuring was done to improve the modularity and readability of the code.